+++ tools/test.py
@@ -400,7 +400,7 @@ if __name__ == "__main__":
             end_frame_id = data_dict['end_frame'].item()
 
             tic = time.time()
-            confmap_pred = rodnet(data.float().cuda())
+            confmap_pred = rodnet(data.bfloat16().cuda())
             if stacked_num is not None:
                 if stacked_num != 1:
                     confmap_pred = confmap_pred[-1].cpu().detach().numpy()  # (1, 4, 32, 128, 128)
